backend/requirements.txt
# Database
psycopg2-binary==2.9.9

# HTTP requests
requests==2.31.0

# HTML parsing (for future implementation)
beautifulsoup4==4.12.2

# HTML parsing
lxml==4.9.3

# Message queue
pika==1.3.2

# Scheduling (optional, if you want more sophisticated scheduling)
schedule==1.2.0

# Testing
pytest==7.4.3
pytest-mock==3.12.0

# Environment variable management
python-dotenv==1.0.0

# For Redis (alternative to RabbitMQ):
#redis==5.0.1

# Logging enhancements (optional)
colorlog==6.7.0

# OCR and image processing
pytesseract==0.3.10
Pillow==10.1.0

# Web API
Flask==2.3.3
Flask-CORS==4.0.0

# Monitoring
prometheus-flask-exporter==0.22.0
